requirements.txt
dash==2.14.2
dash_auth==2.3.0
dash_mantine_components==0.12.0
pandas==2.2.3
plotly==5.22.0
python-dotenv==1.0.1
requests==2.32.2
gunicorn==23.0.0
